+++ src/main/kotlin/com/github/olivernybroe/wingidea/ide/editor/WingPreviewEditorProvider.kt
@@ -15,47 +15,29 @@ import javax.swing.JComponent
 
 
 class WingPreviewEditorProvider: FileEditorProvider {
-    override fun accept(project: Project, virtualFile: VirtualFile): Boolean {
-        return virtualFile.isWingFile
-    }
+    override fun accept(project: Project, virtualFile: VirtualFile): Boolean = virtualFile.isWingFile
 
-    override fun createEditor(project: Project, virtualFile: VirtualFile): FileEditor {
-        // return TextEditorWithPreview(project, virtualFile)
+    override fun createEditor(project: Project, virtualFile: VirtualFile): FileEditor = WingFileEditor(virtualFile)
 
-        return WingFileEditor(project, virtualFile)
-    }
+    override fun getEditorTypeId(): String = "Wing Editor"
 
-    override fun getEditorTypeId(): String {
-        return "Wing Editor";
-    }
-
-    override fun getPolicy(): FileEditorPolicy {
-        return FileEditorPolicy.PLACE_AFTER_DEFAULT_EDITOR
-    }
+    override fun getPolicy(): FileEditorPolicy = FileEditorPolicy.PLACE_AFTER_DEFAULT_EDITOR
 }
 
-class WingFileEditor(project: Project, private val virtualFile: VirtualFile) : FileEditor, UserDataHolderBase() {
+class WingFileEditor(private val virtualFile: VirtualFile) : FileEditor, UserDataHolderBase() {
     private var browser = JBCefBrowserBuilder().setUrl("localhost:3000").build()
 
     override fun dispose() {
         Disposer.dispose(browser)
     }
 
-    override fun getComponent(): JComponent {
-        return browser.component
-    }
+    override fun getComponent(): JComponent = browser.component
 
-    override fun getPreferredFocusedComponent(): JComponent {
-        return browser.component
-    }
+    override fun getPreferredFocusedComponent(): JComponent = browser.component
 
-    override fun getName(): String {
-        return "Wing Preview"
-    }
+    override fun getName(): String = "Wing Preview"
 
-    override fun getFile(): VirtualFile {
-        return virtualFile
-    }
+    override fun getFile(): VirtualFile = virtualFile
 
     override fun setState(state: FileEditorState) {
     }